feat: 로또 생성 로직 분리
