--- src/main/java/lotto/Lotto.java
@@ -1,29 +1,0 @@
-package lotto;
-
-import java.util.Comparator;
-import java.util.List;
-
-public class Lotto {
-    private final List<Integer> numbers;
-
-    public Lotto(List<Integer> numbers) {
-        validate(numbers);
-        this.numbers = numbers;
-        sortNums();
-    }
-
-    private void validate(List<Integer> numbers) {
-        if (numbers.size() != 6) {
-            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
-        }
-    }
-
-    private void sortNums() {
-        numbers.sort(Comparator.naturalOrder());
-    }
-
-    @Override
-    public String toString() {
-        return numbers.toString();
-    }
-}
