+++ src/main/java/lotto/domain/PurchasedLotto.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.Lotto;
-
 import java.util.List;
 
 public class PurchasedLotto {
