+++ src/main/java/lotto/service/LottoService.java
@@ -1,15 +1,12 @@
 package lotto.service;
 
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.domain.PurchasedLotto;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class LottoService {
-    private static final int LOTTO_MIN_NUM = 1;
-    private static final int LOTTO_MAX_NUM = 45;
-    private static final int LOTTO_NUM_PICK_CNT = 6;
 
     private final LottoNumGenerator lottoNumGenerator;
 
@@ -31,11 +28,6 @@ public class LottoService {
     }
 
     private Lotto createLotto() {
-        List<Integer> numbers = generateLottoNumbers();
-        return new Lotto(numbers);
-    }
-
-    private List<Integer> generateLottoNumbers() {
-        return lottoNumGenerator.generateNumbers(LOTTO_MIN_NUM, LOTTO_MAX_NUM, LOTTO_NUM_PICK_CNT);
+        return new Lotto(lottoNumGenerator);
     }
 }
