+++ src/main/java/lotto/view/OutputView.java
@@ -1,7 +1,6 @@
 package lotto.view;
 
-import lotto.Lotto;
-import lotto.domain.PurchasedLotto;
+import lotto.domain.Lotto;
 
 import java.util.List;
